Move webview.dart from platform/ to services/
The platform/ folder name was misleading since it no longer contains
platform-specific code. Moving to services/ aligns better with the
existing code organization (cookie_secure_storage.dart, icon_service.dart,
settings_backup.dart).
